requirements.txt
djangorestframework
pyrebase4
firebase_admin
mysqlclient